pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "dominance-bot"
version = "0.1.0"
description = "Tournament-ready Rocket League bot with hierarchical ML + deterministic control"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "numpy>=1.24",
    "torch>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
rlbot = [
    "rlbot>=1.0",
]
training = [
    "rlgym-sim>=2.0",
    "tensorboard>=2.14",
    "tqdm>=4.66",
]
dev = [
    "pytest>=7.4",
    "black>=23.0",
    "ruff>=0.1",
]
all = ["dominance-bot[rlbot,training,dev]"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
